[vuplus-dvb-proxy] fix postinst error.

+++ meta-bsp/recipes-vuplus/vuplus-dvb-proxy/vuplus-dvb-proxy-vuduo2.bb
@@ -6,9 +6,10 @@ SRCDATE = "20150623"
 SRCDATE_PR = "r0"
 
 pkg_postinst_${PN}_append () {
-  ln -s /lib/modules/${KERNEL_VERSION}/extra/dvb-bcm7424.ko $D/lib/modules/${KERNEL_VERSION}/extra/dvb-bcm.ko
+	if [ ! -f $D/lib/modules/${KERNEL_VERSION}/extra/dvb-bcm.ko ]; then
+		ln -s /lib/modules/${KERNEL_VERSION}/extra/dvb-bcm7424.ko $D/lib/modules/${KERNEL_VERSION}/extra/dvb-bcm.ko
+	fi
 }
-
 
 SRC_URI[md5sum] = "08d5922591f7ea6032d08887de0f9e7d"
 SRC_URI[sha256sum] = "c06c268838f0211d2d8f6608c850c8d0ccc61c7acf02b35c58e317bcc088ab28"
